+++ packages/superdoc/CHANGELOG.md
@@ -1,3 +1,22 @@
+## [0.28.6](https://github.com/Harbour-Enterprises/SuperDoc/compare/v0.28.5...v0.28.6) (2025-11-10)
+
+### Bug Fixes
+
+- additional comments fixes ([#1248](https://github.com/Harbour-Enterprises/SuperDoc/issues/1248)) ([d8d6e52](https://github.com/Harbour-Enterprises/SuperDoc/commit/d8d6e526faa0bd7cd1b7970d3fac5fc8d25f74af))
+
+## [0.28.5](https://github.com/Harbour-Enterprises/SuperDoc/compare/v0.28.4...v0.28.5) (2025-11-10)
+
+### Bug Fixes
+
+- comments not exporting ([e3647a7](https://github.com/Harbour-Enterprises/SuperDoc/commit/e3647a7e3d6a8db7e00f10616b01304e9ca77a39))
+
+## [0.28.4](https://github.com/Harbour-Enterprises/SuperDoc/compare/v0.28.3...v0.28.4) (2025-11-06)
+
+### Bug Fixes
+
+- prettier ([bbc4c6a](https://github.com/Harbour-Enterprises/SuperDoc/commit/bbc4c6a757bcbaca08175dc9bbbe5c533b82e3c4))
+- uploaded image is missing for collaborators ([#1217](https://github.com/Harbour-Enterprises/SuperDoc/issues/1217)) ([090f4e7](https://github.com/Harbour-Enterprises/SuperDoc/commit/090f4e709c560ceae718b34d5a3ee02cc7b2bc86))
+
 ## [0.28.3](https://github.com/Harbour-Enterprises/SuperDoc/compare/v0.28.2...v0.28.3) (2025-11-05)
 
 ### Bug Fixes